# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel API REST - OAuth2 & GIPHY Integration",
        "description": "API REST construida con Laravel 12, OAuth2.0 (Passport), Doctrine ORM y Redis Cache, siguiendo arquitectura hexagonal.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Local Development Server"
        }
    ],
    "paths": {
...
        "/api/v1/media/search": {
            "get": {
                "tags": [
                    "Media"
                ],
                "summary": "Buscar GIFs en GIPHY (con Redis Cache)",
                "description": "Busca GIFs a través de GIPHY API. Las respuestas se cachean en Redis para mejorar performance (Cache Hit: ~5ms, Cache Miss: ~100-300ms).",
                "operationId": "416537e0c59d4e8b0c4730a14067ffaa",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Término de búsqueda (1-50 caracteres)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "funny cats",
                            "maxLength": 50,
                            "minLength": 1
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "description": "Cantidad de resultados (1-50, default: 25)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 10,
                            "default": 25,
                            "maximum": 50,
                            "minimum": 1
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "description": "Offset para paginación (0-4999, default: 0)",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 0,
                            "default": 0,
                            "maximum": 4999,
                            "minimum": 0
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Búsqueda exitosa",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Media encontrado exitosamente"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "string",
                                                        "example": "abc123"
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Funny Cat GIF"
                                                    },
                                                    "url": {
                                                        "type": "string",
                                                        "example": "https://giphy.com/gifs/abc123"
                                                    },
                                                    "rating": {
                                                        "type": "string",
                                                        "example": "g"
                                                    },
                                                    "username": {
                                                        "type": "string",
                                                        "example": "catlovers"
                                                    },
                                                    "images": {
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "pagination": {
                                            "properties": {
                                                "total_count": {
                                                    "type": "integer",
                                                    "example": 1000
                                                },
                                                "count": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "offset": {
                                                    "type": "integer",
                                                    "example": 0
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "meta": {
                                            "properties": {
                                                "status": {
                                                    "type": "integer",
                                                    "example": 200
                                                },
                                                "msg": {
                                                    "type": "string",
                                                    "example": "OK"
                                                },
                                                "response_id": {
                                                    "type": "string",
                                                    "example": "abc123xyz"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Parámetros inválidos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Parámetros inválidos"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "No autenticado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Error de validación",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error de validación"
                                        },
                                        "errors": {
                                            "type": "object",
                                            "example": {
                                                "query": [
                                                    "El parámetro query es requerido"
                                                ]
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "503": {
                        "description": "Error al conectar con GIPHY API",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error al conectar con el proveedor de media"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Error interno del servidor",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": false
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Error interno del servidor"
                                        },
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Ingrese el token JWT en formato: Bearer {token}",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}